Update patch set 3
Patch Set 3: Code-Review-1

Thanks, definitely good to add the additional support.

What will happen if both yaml settings are defined?

I think it would be better to use the ability to distinguish plugin versions to handle this see I834dc05465ea5a24e3743f400837482a82e02e70 for more.

The suggested approach also can't distinguish when there will be a problem trying to upload the multiple version for the older jenkins plugin.

I would suggest converting to expect the list by default, have a compatibility check to distinguish between a string and a list by testing for hasattr(var, __iter__) (strings do not implement this). Output a deprecation warning about the old format and convert it to a list.

Then for the plugin version compatibility, for the older versions if the list len > 1, error out with the information that the plugin is too old. Otherwise generate the necessary XML, and for the newer versions iterate as in the suggested code.

Patch-set: 3
Label: Code-Review=-1