Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Interesting idea, as it will mean included yaml files are truly treated as though they were part of the same file just stored separately.


Suggested alternative approach to avoid needing the extra classes when only overriding one.

I'm guessing that the anchors and aliases are only retained for the current file if it loads additional includes and as soon as the full document is processed from the YamlParser class it will be cleaned out before the next file is processed at the top level.

Would certainly be good if we could add some tests to confirm this and also ensure that if duplicate anchors are defined in the included files that yaml behaves the same as if the yaml code was all in the same file. I think it currently raises an exception.

So that's two tests:

# check that loading two top level files subsequently via the local_yaml.load() method can have the same anchor defined without an issue
# check that loading a file with two includes and the included yaml files contain the same anchor results in an exception.

btw, I believe that PyYaml's behaviour here in raising an exception may not be correct with the current YAML spec, but it probably makes our life easier in ensuring consistent behaviour instead of needing to worry about the load order of the yaml files.

Patch-set: 2
Label: Code-Review=-1