Update patch set 7
Patch Set 7: Code-Review-1

(8 comments)

I think that this needs to raise an exception if the required settings are missing or allow blank values for the current required args.

Seems this plugin normally adds the following XML:

 <overrideAuth>false</overrideAuth>
 <auth>
   <org.jenkinsci.plugins.ParameterizedRemoteTrigger.Auth>
     <NONE>none</NONE>
     <API__TOKEN>apiToken</API__TOKEN>
     <CREDENTIALS__PLUGIN>credentialsPlugin</CREDENTIALS__PLUGIN>
   </org.jenkinsci.plugins.ParameterizedRemoteTrigger.Auth>
 </auth>

While it might be a little too much to try and add support in this patch to configure the different possibilities, perhaps it should include the overrideAuth hardcoded to false to be explicit?


Few minor nits that could be changed at the same time. I think the current code could be made a little more concise without reducing readability.

Patch-set: 7
Label: Code-Review=-1