Update patch set 1
Patch Set 1:

(1 comment)

oops, I can see how that would be annoying.

Since it seems like a good idea to warning the user about obsolete entries, perhaps we should wrap the warning with an additional check so it only outputs a log message if the option is also defined in the yaml and that way it'll be quite once the user has fully converted, complain if they have some of the format and new format, and use the old format otherwise but warn the user that they should switch to using the new format?

typo in the commit message, but ignore it unless changing the patch for another reason ;)

Patch-set: 1
Label: Verified=0