Update patch set 1
Patch Set 1: Code-Review+1

(4 comments)

Given this is a port of existing code I'm tempted to go forward as it is, and submit a separate patch to consolidate and rework the documentation as a separate patch.

Have one lined up but getting the documentation and cross referencing sensible may take a few rounds.

Patch-set: 1
Label: Code-Review=+1