Test command should not attempt connection to Jenkins
By default the test command should skip connecting to a Jenkins
instance. The default configuration will result in using the localhost
to attempt retrieval, any service running against port 8080 on the
localhost or configuration files with non reachable systems specified
will result in a connection error and exception thrown.

The default sample config provided will result in the tool exiting due
to an exception trying to connect to jenkins.example.com when the test
command is used.

Provide a default blank value for plugins_info to skip connecting to a
jenkins instance when using the test command.

Change-Id: I641a3188013debf3765aaff109c4f5fa82a4cedb