Update patch set 2
Patch Set 2:

There's similar change here:
https://review.openstack.org/158084
We should merge that together.

Differences are that 158084 also adds and option to run a DSL script from file. It also always adds the XML element 'buildNeedsWorkspace' to the flow projects, defaulting to false. The reason for that is discussed <a href="https://review.openstack.org/#/c/158084/2/jenkins_jobs/modules/project_flow.py">here</a>.

In this change 150009 are the parameters documented.

Other difference is the option name in yaml - build-needs-workspace vs. needs-workspace.

So, how do we proceed now?

Patch-set: 2