Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Whether we should consider splitting common code like this out into a separate file and module or have it kept with one of the existing entries and simply import it as a function into the other like is done so far probably needs a good answer before we land this.

I think it probably should live somewhere under the modules area though as it isn't common utility code for JJB, it's still plugin configuration code.

Patch-set: 1
Label: Code-Review=-1