Update patch set 2
Patch Set 2: Code-Review-1

This will make it impossible to generate test XML output that matches what will be generated when actually updating a jenkins instance.

The consequence here is that eventually someone (possibly me) will file a bug indicating that JJB is test output does not match the configuration of the job once updated on a Jenkins instance. This will be especially problematic for people (like myself) who use test mode to generate differential output between different commits of their JJB config repository.

If someone other than me runs into this problem, they may begin wasting their time attempting to patch JJB plugin XML generation that generates the wrong XML for a given version of a plugin.

I would note that the behavior this is attempting to "fix" is the same as if attempting to run the "update" subcommand without a "jenkins" stanza in the builder.conf

My suggestion is to add a new option to the builder.conf to switch the proposed new behavior on if someone has a workflow that for whatever reason necessitates running test mode without version information for a specific jenkins.

Patch-set: 2
Label: Code-Review=-1