Update patch set 1
Patch Set 1:

I used common as a placeholder to start this discussion but forgot to specify that when I opened this review. Personally I think the word "common" is a not a good name for such a place so hopefully we can come up with a better one.

I feel like "modules" is not the right place to put this either as when I look at things in modules they all correspond with a JJB section in YAML. I feel this common function does not satisfy that pattern so that's why I created a new folder to organize.

(maybe it doesn't even need a folder and simply should just be a single python file at the upper level)

Patch-set: 1