Update patch set 1
Patch Set 1:

Somewhere under the modules does make a little more sense to me since modules is where the code corresponding to jenkins plugins.

I'll admit the CFP one is a little less common in that it is a plugin that extends the behaviour of other plugins rather than one used directly in job configuration.

So far all these common functions have been left in a module associated with a plugin they come from, and probably they can be left there with only the pure extension style plugins having their configuration placed in a separate file under the modules area.

Or do we want to start moving the common pieces into a another path under the modules folder?

Patch-set: 1