Update patch set 16
Patch Set 16: Workflow-1

I've had time to think about this in more detail with some recent work in depth with yaml and I think it's better if I rewrite the patch so that if you use !include-raw substitution is performed first on the filename and then on the contents, and when you use !include-raw-escape it either skips the substitution or escapes the contents.

Going to investigate if rewriting how the tags are defined to subclass YAMLObject if it allows for more consistent behaviour to be defined in this scenario.

Patch-set: 16
Label: Workflow=-1