Update patch set 9
Patch Set 9: Code-Review+2

(1 comment)

Maybe could include noting that the classpath list requires >=1.6 plugin version in the doc string to be consistent with the others, but I think its release in 2012 is sufficiently long enough ago that it's likely to be the base version for most instances. But if you get a chance to fix it, that would be nice too.

Patch-set: 9
Label: Code-Review=+2