Update patch set 2
Patch Set 2:

Darragh,do you mean we should make the boolean value deprecated and let it be a dict that can contain the key "combine-queued-commits"? In that case, should it be valid to provide an empty dict as parameter (to mean that combine-queued-commits should be false), or should a one element dict be mandatory? If the former, how is the correct yaml syntax to provide an empty dictionary?

Patch-set: 2
Reviewer: Gerrit User 15221 <15221@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0