Update patch set 8
Patch Set 8: Code-Review-1

(4 comments)

Sorry for the delay in getting back reviewing. Looks pretty good to me.

Just a few minor nits around a link in the doc string and where it should use a JenkinsJobException when raising an exception. Also since it's covered both in the pep8 and openstack hacking guidelines, avoid slashes except for last resort and it looks like they can be avoided in this case through use of parenthesis.

Patch-set: 8
Label: Code-Review=-1