Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

Spotted a problem with the literal include.


I've experimented with a few options and it seems that the following is the way to be able to include the docstring from the raw function directly here:

http://stackoverflow.com/questions/7825263/including-docstring-in-sphinx-documentation

Might be possible to get a suitable function implemented to do the needed but would probably take a bit of work to get right.

I'm alright with having the docs directly in the definition in this case with just the remaining issues fixed up. The change to support using it directly from the docstring could be done with a separate change since I think a bit of mucking about with sphinx and how it's subclass would be needed to be done to get this right.

Potentially a similar trick could be used for referencing common options in helper functions to display them in their parents.

Patch-set: 5
Label: Code-Review=-1