Update patch set 8
Patch Set 8:

latest update add the suggested tests.

Was a good thing since revealed that I had added raw into project in setup.cfg but that does not work the same way as the other yaml sections so for now I've removed it.

Keeping this only to be about the modules having raw support.

Patch-set: 8
Label: Verified=0