Update patch set 19
Patch Set 19:

My mac was using a wifi connection which is probably not a good idea when running these tests.  I plan to retry with a wired connection next time i have the opportunity.   This time I ran the tests on my Linux VM and here are the results.


I was able to repro the the error "ValueError: too many values to unpack" when I ran the command "jenkins-jobs --conf jjb_config.ini update ../project-config/jenkins/jobs".  I was able to repro when running right on the jenkins master and on a remote client.  I thought this might be an existing problem so I tried running it from the master branch but was not able to repro it there.

  More info on my config here: http://paste.openstack.org/show/199650/

The good news is that when I run "jenkins-jobs --conf jjb_config.ini update --workers 8 ../project-config/jenkins/jobs" I don't get any errors at all either from local or remote clients.

Patch-set: 19