Use default empty plugins list for all Jenkins exceptions
With the existing implementation it was not possible to run the test
command without having a live Jenkins connection, or using credentials
that does not provide access to the /pluginManager endpoint.

This causes problems for users who run JJB against job configurations
only to test the yaml syntax and don't care about the plugin versions
on the server.

Remove the condition in get_plugins_info's exception handling and use
the default empty plugin list on all Jenkins failures.

Users who care about plugin versions can still specify a plugin-info
file.

Change-Id: I080467bbcf71810828e5d765f93f113bb441467f