Support default string values for variables
Provide syntax support for specifying default values to be substituted
for variables during deep_format when no other replacement is provided.

Allows for individual variables to have a default or be optionally blank
should nothing be placed after the custom specifier.

To allow for default variables to be used in template names without
needing to include the syntax with the default value, add support to
reference any block through an optional 'id'. Use multi_key_dict to
allow reference via both 'name' and 'id'.

Change-Id: Ib97a33a2bbca123791d4ca6ef5248ed200992565