Enable parameter ordering in trigger_parameterized_builds
This is necessary to to allow users to explicitly order the
"current-parameters", "property-file", and "predefined-parameters" when
triggering parameterized jobs. The XML ordering matters when Jenkins is actually
executing the Trigger Parameterized Builds plugin in cases where these various
sources attempt to define the same parameter.

ie if FOO=BAR is set as a build parameter of the current job and a properties
file is generated during build that attempts to set FOO=BAZ; the XML ordering
determines which value, BAZ or BAR, that the triggered job sees for the
parameter FOO.

Change-Id: I13716dba384b7134108cec2c31668c5069564205