Update patch set 4
Patch Set 4:

(1 comment)

It's not a bug in the plugin. In fact, I consider it a feature of the plugin that the order of XML elements actually matters--otherwise we'd be forced to use whatever the Trigger Parameterized Builds plugin author thinks should be the override order.

When dealing with these configuration options in the Jenkins Web UI, the order in which you "Add Parameters" is what determines the XML configuration on disk (which in turn affects the behavior of the plugin). So as far as writing the XML to disk, the Trigger Parameterized Build plugin has less to do with the ordering than the generic web ui interface for adding configuration blocks.

Patch-set: 4
Label: Verified=0