Add option to skip certification check
adds a `no_check_certificate` to the [jenkins] section.

If set to True https/ssl connections will skip certificate check,
allowing you to connect to systems that might have an outdated or
non-public certificate. Default is and should always be False.

Supported only by Python 2.7.9+ below it will be ignored.

Change-Id: Ieca1c07f5142b7ac379215bfed35ca8c9d8b771e