Adjust precedence levels of params
Should always allow params specified for jobs/job-groups to override the
defaults set on these definitions.

The precedence order was project -> job -> group -> group-job, with the
latter ones taking precedence over the former.

Simple project precedence:

    project params -> job_list -> jobs params

Job group precedence:

    group params -> job_list -> group jobs params

However when substituting the job_list in a project with a job-group to
gather together defaults the previous precedence level resulted in:

    project params -> jobs params -> (
        group params -> job_list -> group jobs params)

When it would appear to make more sense to have:

    project params -> (
        group params -> job_list -> group jobs params
	    ) -> jobs params

Change-Id: I70a0eb63b0e129aaa46f3a8b4393d7ac15fa9e84