Ensure tests run with ConfigParser object
Should always use a ConfigParser object when testing to allow unit tests
to check that calls behave as expected.

Read in the default configuration to align with expected usage, and fix
some issues caused by usage of dicts resulting in the option name
requested being taken as the default value.

Since '.get()' is a valid method for a dictionary object, take care to
ensure that we do not use a dummy dict object as this will not throw the
expected errors if a config file is not provided and will also treat the
option name as a default value resulting in the wrong data appearing in
the outputted XML.

Change-Id: I6abb85f0363a48f220914c52b74e5a5fb65b0faa