Update patch set 2
Patch Set 2:

Thought about this a bit more and I think the that we should just stream the multiple yaml files through to the loader and leave it up to the user to keep everything in the multiple files in the correct format to work together. So if they mix a list with a dict, the yaml loader will throw an error.

Patch-set: 2