Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Some suggestions on changing how the names are being cached and queries on whether to consider caching the job objects as well.

At the very least I'd suggest switching to using a set() for storing the job names, especially with large numbers of lookups it'll be much more efficient.

Patch-set: 3
Label: Code-Review=-1