Update patch set 1
Patch Set 1:

If I was calling 'update_job()' programatically, i.e. treating JJB as an API, if would be really useful if it returned some information that clearly stated what it had changed. Removing the list is fine for users reading the log output, but removes something that is sensible for a program to consume.

Patch-set: 1