Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Looks good, that should definitely cut down on the number of API queries!

Just a small tweak on invalidating the cache would ensure both sets of cached data remain consistent.

Be interested in knowing the sort of numbers around the performance improvement here. Obviously it'll be a improvement in removing an API query per job, but knowing the real world impact would be a nice FYI :)

Patch-set: 6
Label: Code-Review=-1