Update patch set 7
Patch Set 7:

To put some context around the improvement in performance in case people wonder how it could cause such a big impact.

When querying the list of jobs in jenkins, jenkins must query internally it's only list of jobs, which if it is not being cached very intelligently any time a job is updated or created it's likely Jenkins' internal cache is invalidated. (note to self, go look at the jenkins code to determine this).

Taking some simple variables:
x = time to query a single job internally to jenkins
y = time take to make an API query

If you have 100 jobs, the time take to respond is:

 100x + y

If you query this 100 times, the worst case scenario is:

 100(100x + y)

More likely this is, if P(c) is probability of a cache hit internal to Jenkins:

 100(100(1 - P(c))x + y)

or more generically:

 n(n(1 - P(c))x + y)

Which will contain an n^2 statement, thus the more jobs you have the significantly worse the problem gets since it's not linear, it's quadratic.

With this change, this now becomes:

 n(1 - P(c))x + y

Which is O(n) based and likely means that the benefits/costs of how well the cache in Jenkins is negligible and so becomes:

 nx + y

But it should be very obvious that for larger n:

 nx + y << n^2x - n^2P(c))x + ny

Patch-set: 7