Update patch set 1
Patch Set 1:

(1 comment)

I'll try to get to the documentation this weekend.

The general use case is to when using job-templates in order to make them highly reusable there are many variables added, the hitch is that this means you need to define a value for these variables for all uses of the job template even if the majority use the same setting and it's only a few additional jobs that need to use something else.

The alternative is a separate job-template or multiple job-groups.

Essentially without a mechanism to provide for default values you end up creating more yaml definitions or expanding on the existing ones with a lot of additionally boiler plate just to deal with a handful of edge cases.


I think this change is still incomplete, as some consideration about how template names are affected by setting such variables needs to be done. I'm tempted to suggest that each template should support an 'id' attribute, which is the 'name' by default, but can be overridden by users explicitly setting it.

Then when working out the 'id' if it's not set by the yaml configuration, the parser would process the name to remove the custom syntax placeholder for defaults so that the template name may include variables with default values without anyone needing to update their references to such templates.

Patch-set: 1