Refactor err reporting of incorrect YAML structure
This improves the error reporting of incorrect structure of the YAML -- when
the missing indent is detected. It adds the information about the file
name with problematic structure.

Code which is extracting the expected single item from a collection has
been extracted to `get_the_only_one_subitem` function to increase the
readability.

Added tests for missing indent exception.

Change-Id: Ieebc778416998f3c3427f957c942391f2f5475dd