Update patch set 3
Patch Set 3:

One concern I have with this is that in theory yaml is just fine loading a utf16 encoded file and this change firmly puts JJB in a position where that ins't supported.

I don't think this is worth a -1 because we don't test the utf16 case and I expect it to not work for other reasons (because we have assumed utf8 for the most part). But it is worth pointing out that this would need to be reverted or at least rewritten if we were to ever try and support not utf8 for yaml files.

Can anyone come up with a reason for why this would be worth a -1? If not I will properly review and likely approve.

Patch-set: 3
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0