Update patch set 2
Patch Set 2:

I won't +/- since this is my first time speaking up in a review, but I would like to second Wayne Warren's concerns.

Both about telegraphing broadly and explicitly backwards breaking changes and that if we're changing the precedence order we should go further.

Conceptually we use defaults for all or a class of jobs, job templates that are then grouped into different pipelines (job-groups) which are then instantiated per project.

Understanding that JJB does not interpolate in that order has cause a considerable amount of confusion for members of our team and we have significant work arounds in place because of it.

Also, if there's a better way to model those concepts then using those constructs in that way, I would appreciate references to documentation. It is my team's understanding that this precedence ordering is out of of date to the constructs now in JJB, and this patch gets us *most* of the way to how we originally believed the precedence would work.

Patch-set: 2
Reviewer: Gerrit User 16480 <16480@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0