Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Needs a tweak to be python 3 compatible. Haven't looked too closely otherwise. Not sure how useful it is combining this exception with the YamlFormatError as it seems that the only place that would have the file name is during the parse unless you can think of a way to save that information with each item to be retrieved when exceptions occur?

Patch-set: 2
Label: Code-Review=-1