Update patch set 1
Patch Set 1: Code-Review+1

Proposed changes make sense to me. In general I believe a Python list already has some guarentees that the order they are added is the same order that they will be returned, so I'm not sure you need to additionally test the list() itself.

Maybe instead a test on _find_file() to prove that if the same file exists in 2 paths that the higher listed path takes precedence would be sufficient?

Patch-set: 1
Label: Code-Review=+1