Update patch set 1
Patch Set 1: Code-Review-1

I think this is unnecessary, perhaps a documentation update would be better. This is not documented but commenting out the username and password field in the config file works for anonymous setup on a Jenkins instance that's configured without security. I use this often with the official Jenkins docker for testing.

Patch-set: 1