Update patch set 1
Patch Set 1: Workflow-1

The current default is not useful, the odds that Jenkins would ever accept empty strings as a username password combination is very low. This at least means that the default behavior will work with a default Jenkins.

Needing to write a config file for an unsecured Jenkins really doesn't make sense.

Patch-set: 1