Update patch set 6
Patch Set 6: Code-Review-1

(7 comments)

I think there is a bit of tidy up on the code that can be done. Seems like there is some duplication that removed, and some code can be moved outside of the loop as well. Few things fall out from that but otherwise looks good.

Minor nit on the docstring style, as I've suggested some code clean up I think it's worth trying to adhere to the more concise style (we'll probably want to consolidate all the docstrings on that one to make them easeir to read and shorter ;) )

Patch-set: 6
Label: Code-Review=-1