Update patch set 13
Patch Set 13: Code-Review-1

I just updated the syntax of the commit message in this PS.

I think this could be really useful and am in favor of it generally.  I also think we could probably take advantage of it in OpenStack.

My main objection to this is that there is no way to override defaults, particularly, say, to remove a wrapper defined in defaults.  I think that is an important thing to be able to support.  I gather from IRC that Richard has a follow-on patch to allow that.  It might be good to review that along with this to make sure that the whole package makes sense.  But assuming it does, that, along with having this as a non-default configurable behavior should address my concern.

I think it would be good to include documentation in this patch as well (even if that documentation is modified in the later patch that adds override support).

Finally, I did have to think rather hard about what the desired effect of the deep merge operations was.  The intent of this is not entirely understandable without the context of the bug report.  The documentation I mentioned above should help a bit, but also a couple of code comments would be useful too.

Thanks!

Patch-set: 13
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1