Update patch set 3
Patch Set 3:

Currently if we tried to open a UTF-16 input and read it, we don't convert correctly to the utf-8 that we are uploading to Jenkins so it couldn't possibly work.

I think that if we want to support multiple input codecs we need to make use of either https://pypi.python.org/pypi/chardet or http://www.crummy.com/software/BeautifulSoup/bs3/documentation.html#Beautiful%20Soup%20Gives%20You%20Unicode,%20Dammit (which uses chardet).

I think we're currently in no worse a position than before for UTF-16, and argubly in a better place to enhance to support handling it in the future as an input via detection and conversion to utf-8.

Patch-set: 3
Label: Verified=0