Update patch set 14
Patch Set 14: Code-Review-1

This is currently broken, if you review the compare build output from patchset 11 and grep for gate-magnetodb-devstack-dsvm in the output:

http://logs.openstack.org/30/150730/11/check/jenkins-job-builder-compare-xml/1f5191d/console.html.gz

you'll see a change where the timeout goes from 65 minutes to 30 minutes.

However you can see from the config entry for this job, it is setting timeout to 65 minutes:
https://github.com/openstack-infra/project-config/blob/master/jenkins/jobs/magnetodb.yaml#L7


The code is overriding the job specific settings with those from the defaults. This is not correct, the deep merge should only merge in defaults where they are not already set.

I would recommend that the config option be inverted for now and the workflow be set to -1 until the code is tweaked so only those jobs that are missing settings defined by the defaults receive them when deep_merge is enabled and also that a test be added that shows the correct behaviour when some of the values are already set on a job.

Patch-set: 14
Label: Code-Review=-1