Update patch set 15
Patch Set 15: Code-Review-1

I've identified the main problem as being how macros are being expanded/deep-merged and tweaked the test to expose the issue that was also impacting the infra job configs.

Once that is fixed, I think it'll be safe to merge and likely will work as expected as well.

Patch-set: 15
Label: Code-Review=-1