Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

Just a few minor nits. I guess if the folder name is embedded in the job name, it should mean that the alphabetical ordering will ensure the golder is created before any jobs contained.

However it may be useful to handle folders in a special way in the future to ensure they are iterated over first to allow for some parallel behaviour when uploading jobs.

Nothing to be concerned about here though I think.

Patch-set: 1
Label: Code-Review=-1