Update patch set 2
Patch Set 2: Code-Review-1

Looks like the purpose of this change is to have an ordered list of paths so unless i'm missing something i don't see how the tests in this patchset verify that functionality.   How about if we do the path ordering verification using the existing multi tests in test_test.py?

Patch-set: 2