Update patch set 1
Patch Set 1: Workflow-1

Took me a while to remember the exact case for this. Jenkins supports anonymous read access at the same time as requiring a user/password combination to perform actions.

While running in test mode you can point to a real jenkins instance in order to query the plugins and use the versions given to control the XML generated. To do this as an anonymous user you must have both user & password set to None to prevent the python-jenkins library trying to set the authentication header which would break the query for the plugins.

Apologies it's been so long since I wrote this and I forgot myself what exactly it was for. I need to tweek both the comments and commit message to make this clear.

Patch-set: 1
Label: Workflow=-1