Update patch set 16
Patch Set 16: Code-Review-1

(2 comments)

Some minor nits on the return statements.

I think this can be cleaned up a lot.

Should probably just look to see if the attribute is a collection of some sort i.e. has __iter__ but not format as methods. This would distinguish between basic types, instead of trying to detect each of the basic types in turn and then looking for lists/sets and finally dictionaries.

Right now, I suspect whether works correctly depends on the data type being used, so a float would probably cause the same issue that the use of the integer for the timeout did.

I need to do some experimentation to see just how much can be removed and refactored, but definitely as it stands since it's not distinguishing correctly between collections and basic types it shouldn't go in.

Patch-set: 16
Label: Code-Review=-1