Update patch set 6
Patch Set 6:

(3 comments)

I'm mildly concerned that we're papering over a more generic problem in a manner that we'll have to keep amending the code to deal with additional parameter ordering issues as well as tackle the same issue elsewhere.

One of the primary drivers for use of OrderedDict was that we realised the order of the output in many cases was important and should ideally match the input order of that yaml so that it was user controllable.

Seems this is another case where we should be leveraging that ordered structure of the data read from the yaml, though I understand the desire to use a separate element to avoid breaking backwards compatibility it does seem like using the input order would be the best way forward and I'm not sure just how tricky it would be support the different mechanisms in the future if this change lands as is.

Thoughts?

Patch-set: 6
Label: Verified=0