Update patch set 3
Patch Set 3: Code-Review+1

No problem. The comments are attached to a specific patch so if you click on a reviewed file that has a comment you additionally need to click the patch which the comment was added to from the top options that say "Base 1 2 3 4 ..."

Hope this helps.

Patch-set: 3
Label: Code-Review=+1