add support for Image Gallery publisher plugin
You can find the original version of this patch here:
  https://github.com/h01ger/jenkins-job-builder/commit/5a04d4b40523fb021ef2d9b2b6e508f35dfbe027

My (Philip Hands's) contribution is limited to moving the one line from
setup.py to setup.cfg, fixing a couple of very minor pep8 errors, adding
the unit test and submitting it upstream (now).
I've therefore set the Author to be Holger.

It is being submitted upstream by me (Philip Hands) with Holger's approval,
because he doesn't have the time available to do it himself.

Note that the code includes a couple of FIXME comments -- I'm not really
familiar with jenkins, or j-j-b ... otherwise I'd have attempted to
address the issues that these comments highlight.  If these need to be
fixed before the code would be deemed acceptable, and if explaining to
me what is involved in doing that seems likely to be less effort that
just doing it yourself, feel free to give that a try :-)

Change-Id: Ie3be52643e40d827df935cf18eca41de9589987f
Co-Authored-By: Philip Hands <phil@hands.com>