scm svn: add support for viewvc-url
This allows one to specify a URL that links to the web interface for the source being built.

You can find the original version of this patch here:

  https://github.com/h01ger/jenkins-job-builder/commit/e38c41ef6e116072d9965f13cbb9d55eebe51f57

My (Philip Hands's) contribution is limited to fixing a very minor pep8
error, adding feature to a unit test, and submitting it upstream (now).
I've therefore set the Author to be Holger.

It is being submitted upstream by me (Philip Hands) with Holger's approval,
because he doesn't have the time available to do it himself.

Change-Id: Ia5acadc5b7bb1a144cc375a692cb7ad0474a931f
Co-Authored-By: Philip Hands <phil@hands.com>