Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

Sorry about the delay in getting back to review this. The hipchat exception should really be MissingAttributeError as the data being read in is no longer yaml, but python data structures that may or may not have been represented in yaml.

Also I think it makes more sense to rename the group of tests yaml_structure_error rather than just missing indent since it's also possible to trigger the exception in including additional unexpected entries which are not indented incorrectly but should not be there.

Patch-set: 3
Label: Code-Review=-1