Update patch set 2
Patch Set 2:

I've added a test case that uses macros and a job template and it retains the ordering using the yaml provided the necessary config option is turned on.

Could you provide an example of where this isn't working as expected as I'm not sure what didn't work.

Patch-set: 2
Label: Verified=0