Update patch set 2
Patch Set 2:

Hadn't spotted the py34 failure. The reason that the py27 tests are passing is partially that PYTHONHASHSEED=0, which is disabling any random dict behaviour resulting in consistent return order. This just happens to match the desired behaviour, while for py34 the consistent return order for dict doesn't.

Patch-set: 2