Update patch set 7
Patch Set 7: Code-Review+1

I added the other patch to address this issue.

I raised this as a bug here: https://storyboard.openstack.org/#!/story/2000301, though I wasn't sure what the underlying cause of the problem was. Is it confirmed that it's due to moving to Hipchat plugin 0.1.9?

This seems to address the problem I was having, so it's a +1 from me. I'm not terribly familiar with the JJB code base, so I think you ought to wait for other approval than mine.

In the mean time, I'll close the other pull request, as having multiple open to address the same issue is confusing.

Patch-set: 7
Label: Code-Review=+1