Add support to reverse trigger for job list
Listing jobs as a comma separated list is fine with small amounts of
jobs however when this list gets too long it can be difficult to
maintain.

This patch adds support to list jobs as a list and the code will
automatically change it to a python separated list. Also retains support
for the old style comma separated list.

Change-Id: I53e9ddb312d2588ad63249714da11e7bf35cc960
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>