Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

It's about the wording used in the doc string for the 'publishers', it current states it's the list of projects to trigger, when it should say it's the list of publishers to trigger. Might just be a little confusing.

FYI you can also switch which patchset is used as the base comparison set when doing side-by-side comparisons to see previous comments for a quick check against the latest state.

Patch-set: 5
Label: Code-Review=-1